+++ lib/puppet/provider/wls_dynamic_cluster/simple.rb
@@ -1,7 +1,7 @@
 require 'easy_type'
 require 'utils/wls_access'
 
-Puppet::Type.type(:wls_domain_partition).provide(:simple) do
+Puppet::Type.type(:wls_dynamic_cluster).provide(:simple) do
   include EasyType::Provider
 
   desc 'Manage all the WebLogic 12.2.1 domain partitions via regular WLST'